Moved security configuration to main
